Extract TreeSearchInput to separate component
